Create DragAndDrop
posting drag and drop script. This script can be saved as a .bat file on a Windows system, and used to create abridge MediaConch policy checker reports as text files, which just state whether the file passed or failed. Dependent on MediaConch, XMLStarlet, and an existing policy. This script uses the "LPB_AccessCopySD.xml" policy.